added key delete

+++ key-client/src/key/gui/EditComposite.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.util.Collection;
 
 import key.model.CsvExport;
-import key.model.DBLayer;
 import key.model.KeyData;
 import key.model.KeyException;
 import key.model.SetData;
@@ -19,6 +18,7 @@ import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.FileDialog;
 import org.eclipse.swt.widgets.Label;
+import org.eclipse.swt.widgets.MessageBox;
 
 public class EditComposite extends Composite {
 
@@ -26,8 +26,10 @@ public class EditComposite extends Composite {
 	private SetTable tableSet;
 	private KeyTable tableKey;
 	private Button buttonDeleteSet;
+	private Button buttonDeleteKey;
 
 	private SetData selectedSetData;
+	private KeyData selectedKeyData;
 
 	public EditComposite(Composite parent, int style) {
 		super(parent, style);
@@ -78,8 +80,16 @@ public class EditComposite extends Composite {
 				buttonDeleteSet.setEnabled(keys != null);
 			}
 		});
+
 		tableKey = new KeyTable(this);
 		tableKey.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 2, 2));
+		tableKey.addSelectionListener(new KeyTable.KeySelectionListener() {
+			@Override
+			public void selected(KeyData keyData) {
+				selectedKeyData = keyData;
+				buttonDeleteKey.setEnabled(keyData != null);
+			}
+		});
 
 		buttonDeleteSet = new Button(this, SWT.PUSH);
 		buttonDeleteSet.setLayoutData(buttonGridData);
@@ -88,27 +98,27 @@ public class EditComposite extends Composite {
 		buttonDeleteSet.addSelectionListener(new SelectionAdapter() {
 			@Override
 			public void widgetSelected(SelectionEvent e) {
-				deleteSet(selectedSetData);
+				deleteSelectedSet();
 			}
 		});
 		buttonDeleteSet.setEnabled(false);
 
-		button = new Button(this, SWT.PUSH);
-		button.setText("Удаление ключа");
-		button.setLayoutData(buttonGridData);
-		button.setImage(new Image(getShell().getDisplay(), getClass().getResourceAsStream("/delete.gif")));
-		button.addSelectionListener(new SelectionAdapter() {
+		buttonDeleteKey = new Button(this, SWT.PUSH);
+		buttonDeleteKey.setText("Удаление ключа");
+		buttonDeleteKey.setLayoutData(buttonGridData);
+		buttonDeleteKey.setImage(new Image(getShell().getDisplay(), getClass().getResourceAsStream("/delete.gif")));
+		buttonDeleteKey.addSelectionListener(new SelectionAdapter() {
 			@Override
 			public void widgetSelected(SelectionEvent e) {
-				deleteKey();
+				deleteSelectedKey();
 			}
 		});
-		button.setEnabled(false);
+		buttonDeleteKey.setEnabled(false);
 	}
 
 	private void refresh() {
 		tableSet.reset();
-		for(SetData setData : DBLayer.load()) {
+		for(SetData setData : SetData.listAll()) {
 			tableSet.setValues(setData);
 		}
 	}
@@ -127,13 +137,24 @@ public class EditComposite extends Composite {
         }
 	}
 
-	private void deleteSet(SetData setData) {
-		DBLayer.delete(setData);
-		tableSet.deleteCurrent();
+	private void deleteSelectedSet() {
+		MessageBox mb = new MessageBox(getShell(), SWT.YES | SWT.NO);
+		mb.setText("Удаление пачки");
+		mb.setMessage("Вы уверены?");
+		if(SWT.YES == mb.open()) {
+			selectedSetData.delete();
+			tableSet.deleteCurrent();
+		}
 	}
 
-	private void deleteKey() {
-//		tableKey;
+	private void deleteSelectedKey() {
+		MessageBox mb = new MessageBox(getShell(), SWT.YES | SWT.NO);
+		mb.setText("Удаление ключа");
+		mb.setMessage("Вы уверены?");
+		if(SWT.YES == mb.open()) {
+			selectedSetData.deleteKey(selectedKeyData);
+			tableKey.deleteCurrent();
+		}
 	}
 
 }
